# src/api/schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Healthcare Simulation API",
    "description": "Healthcare simulation API with integrated Ollama LLM capabilities",
    "version": "0.3.0",
    "contact": {
      "name": "Healthcare Simulation Team",
      "url": "https://healthcare-simulation-api.onrender.com"
    }
  },
  "servers": [
    {
      "url": "https://localhost",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/v1/healthcare/simulate": {
      "post": {
        "summary": "Simulate healthcare scenario",
        "description": "Process a healthcare simulation scenario using advanced medical LLM model",
        "operationId": "simulate_healthcare_scenario",
        "tags": ["Healthcare Simulation"],
        "security": [{"APIKeyHeader": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful simulation response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResponse"
                }
              }
            }
          },
          "401": {"$ref": "#/components/responses/UnauthorizedError"},
          "500": {"$ref": "#/components/responses/ServerError"}
        }
      }
    },
    "/v1/ollama/generate": {
      "post": {
        "summary": "Generate medical response",
        "description": "Generate medical response using Ollama model",
        "operationId": "generate_response",
        "tags": ["Ollama Integration"],
        "security": [{"APIKeyHeader": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              }
            }
          },
          "401": {"$ref": "#/components/responses/UnauthorizedError"},
          "500": {"$ref": "#/components/responses/ServerError"}
        }
      }
    },
    "/v1/ollama/chat": {
      "post": {
        "summary": "Interactive medical chat",
        "description": "Start or continue a medical chat session",
        "operationId": "chat",
        "tags": ["Ollama Integration"],
        "security": [{"APIKeyHeader": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "401": {"$ref": "#/components/responses/UnauthorizedError"},
          "500": {"$ref": "#/components/responses/ServerError"}
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SimulationRequest": {
        "type": "object",
        "required": ["scenario", "title", "actors"],
        "properties": {
          "scenario": {
            "type": "string",
            "description": "Healthcare scenario to simulate"
          },
          "title": {
            "type": "string",
            "description": "Title of the simulation scenario"
          },
          "actors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of actors involved in the scenario"
          },
          "patient_data": {
            "$ref": "#/components/schemas/PatientData"
          }
        }
      },
      "SimulationResponse": {
        "type": "object",
        "properties": {
          "diagnosis": {
            "type": "string",
            "description": "Preliminary diagnosis"
          },
          "recommended_actions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "vital_signs": {
            "$ref": "#/components/schemas/VitalSigns"
          },
          "risk_assessment": {
            "type": "string"
          }
        }
      },
      "GenerateRequest": {
        "type": "object",
        "required": ["model", "prompt"],
        "properties": {
          "model": {
            "type": "string",
            "description": "Name of the Ollama model"
          },
          "prompt": {
            "type": "string",
            "description": "Medical prompt"
          },
          "options": {
            "$ref": "#/components/schemas/GenerationOptions"
          }
        }
      },
      "GenerateResponse": {
        "type": "object",
        "properties": {
          "response": {
            "type": "string",
            "description": "Generated medical response"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "model": {
                "type": "string"
              },
              "eval_duration": {
                "type": "integer"
              }
            }
          }
        }
      },
      "ChatRequest": {
        "type": "object",
        "required": ["model", "messages"],
        "properties": {
          "model": {
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            }
          }
        }
      },
      "ChatMessage": {
        "type": "object",
        "required": ["role", "content"],
        "properties": {
          "role": {
            "type": "string",
            "enum": ["system", "user", "assistant"]
          },
          "content": {
            "type": "string"
          }
        }
      },
      "ChatResponse": {
        "type": "object",
        "properties": {
          "message": {
            "$ref": "#/components/schemas/ChatMessage"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "model": {
                "type": "string"
              },
              "eval_duration": {
                "type": "integer"
              }
            }
          }
        }
      },
      "GenerationOptions": {
        "type": "object",
        "properties": {
          "temperature": {
            "type": "number",
            "minimum": 0,
            "maximum": 2
          },
          "top_p": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          }
        }
      },
      "PatientData": {
        "type": "object",
        "properties": {
          "age": {
            "type": "integer"
          },
          "gender": {
            "type": "string",
            "enum": ["male", "female", "other"]
          },
          "vital_signs": {
            "$ref": "#/components/schemas/VitalSigns"
          }
        }
      },
      "VitalSigns": {
        "type": "object",
        "properties": {
          "heart_rate": {
            "type": "integer"
          },
          "blood_pressure": {
            "type": "object",
            "properties": {
              "systolic": {
                "type": "integer"
              },
              "diastolic": {
                "type": "integer"
              }
            }
          },
          "respiratory_rate": {
            "type": "integer"
          },
          "temperature": {
            "type": "number"
          },
          "oxygen_saturation": {
            "type": "integer"
          }
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Invalid API key",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}